requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
python-multipart==0.0.6
python-dotenv==1.0.0
langchain-core==0.3.46
langchain-community==0.3.20
langchain-openai==0.3.9
langchain-text-splitters>=0.3.0
langchain-unstructured==0.1.5
unstructured-inference==0.8.10
unstructured[pdf]==0.17.0
python-docx>=1.1.2
openai>=1.66.3,<2.0.0
pypdf>=3.0.0
pytesseract==0.3.10
pdf2image==1.16.3
Pillow==11.1.0
pdfminer.six>=20221105
docx2txt>=0.8
protobuf==6.30.0
onnx>=1.15.0
onnxruntime==1.21.0